一旦機能実装完了

+++ src/main/java/jp/co/monocrea/controller/UserController.java
@@ -1,8 +1,7 @@
 package jp.co.monocrea.controller;
 
-import java.util.List;
-
 import jakarta.inject.Inject;
+import jakarta.validation.Valid;
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.DELETE;
 import jakarta.ws.rs.GET;
@@ -14,7 +13,8 @@ import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.QueryParam;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
-import jp.co.monocrea.dto.SearchUserResult;
+import jp.co.monocrea.dto.UserSearchResult;
+import jp.co.monocrea.dto.UserCreateRequest;
 import jp.co.monocrea.entity.User;
 import jp.co.monocrea.enums.OrderType;
 import jp.co.monocrea.repository.UserRepository.SortItem;
@@ -33,14 +33,14 @@ public class UserController {
     public Response searchUser(
             @QueryParam("id") Integer id,
             @QueryParam("userName") String userName,
-            @QueryParam("_sort") String sortItemName,
+            @QueryParam("_sort") String sort,
             @QueryParam("_order") String order,
             @QueryParam("_page") Integer page,
             @QueryParam("_limit") Integer limit
             ){
         try {
-            if (sortItemName == null) {
-                sortItemName = SortItem.ID.getValue(); // デフォルトはユーザＩＤ
+            if (sort == null) {
+                sort = SortItem.ID.getValue(); // デフォルトはユーザＩＤ
             }
 
             if(order == null){
@@ -55,11 +55,11 @@ public class UserController {
                 limit = 10; // デフォルトの取得件数は10
             }
 
-            SearchUserResult result = userService.searchUser(id, userName, sortItemName, order, limit, page);
+            UserSearchResult result = userService.searchUser(id, userName, sort, order, limit, page);
             return Response.ok(result).build();
         } catch (Exception e) {
             return Response.status(Response.Status.INTERNAL_SERVER_ERROR)
-                           .entity("ユーザの検索に失敗しました。: " + e.getMessage()) // TODO: エラーメッセージのローカライズ
+                           .entity("ユーザの検索に失敗しました。: " + e.getMessage())  // TODO: エラー時レスポンス構成は要検討
                            .build();
         }
     }
@@ -67,17 +67,15 @@ public class UserController {
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public Response createUser(User user) {
-        // TODO: リクエストパラメータのバリデーションチェック
-
+    public Response createUser(@Valid UserCreateRequest request) {
         try {
-            user = userService.createUser(user.getUserName(), user.getUserNameKana());
+            User user = userService.createUser(request.getUserName(), request.getUserNameKana());
             return Response.status(Response.Status.CREATED)
                            .entity(user)
                            .build();
         } catch (Exception e) {
             return Response.status(Response.Status.INTERNAL_SERVER_ERROR)
-                           .entity("ユーザの作成に失敗しました。: " + e.getMessage()) // TODO: エラーメッセージのローカライズ
+                           .entity("ユーザの作成に失敗しました。: " + e.getMessage())  // TODO: エラー時レスポンス構成は要検討
                            .build();
         }
     }
@@ -86,14 +84,24 @@ public class UserController {
     @Path("/{id}")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public Response updateUser(@PathParam("id") Integer id, User user) {
+    public Response updateUser(@PathParam("id") Integer id, @Valid UserCreateRequest request) {
+        if( id == null || id <= 0) {
+            return Response.status(Response.Status.BAD_REQUEST)
+            .entity("ユーザIDが不正です。") // TODO: エラー時レスポンス構成は要検討
+            .build();
+        }
+
         try {
+            User user = new User();
             user.setId(id); 
+            user.setUserName(request.getUserName());
+            user.setUserNameKana(request.getUserNameKana());
+
             user = userService.updateUser(user);
             return Response.ok(user).build();
         } catch (Exception e) {
             return Response.status(Response.Status.INTERNAL_SERVER_ERROR)
-                           .entity("ユーザの更新に失敗しました。: " + e.getMessage()) // TODO: エラーメッセージのローカライズ
+                           .entity("ユーザの更新に失敗しました。: " + e.getMessage())  // TODO: エラー時レスポンス構成は要検討
                            .build();
         }
     }
@@ -103,12 +111,18 @@ public class UserController {
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
     public Response deleteUser(@PathParam("id") Integer id) {
+        if (id == null || id <= 0) {
+            return Response.status(Response.Status.BAD_REQUEST)
+                           .entity("ユーザIDが不正です。") // TODO: エラー時レスポンス構成は要検討
+                           .build();
+        }
+
         try {
             userService.deleteUser(id);
             return Response.noContent().build();
         } catch (Exception e) {
             return Response.status(Response.Status.INTERNAL_SERVER_ERROR)
-                           .entity("ユーザの削除に失敗しました。: " + e.getMessage()) // TODO: エラーメッセージのローカライズ
+                           .entity("ユーザの削除に失敗しました。: " + e.getMessage())  // TODO: エラー時レスポンス構成は要検討
                            .build();
         }
     }
